requirements.txt
fastapi
uvicorn
nemo_toolkit[all]
onnxruntime
torchaudio
soundfile
numpy